+++ tests/devtools-evm-hardhat-test/hardhat.config.with-valid-and-invalid-rpcs.ts
@@ -24,7 +24,7 @@ const config: HardhatUserConfig = {
         },
         arbSepolia: {
             eid: EndpointId.ARBSEP_V2_TESTNET,
-            url: 'https://arbitrum-sepolia.blockpi.network/v1/rpc/public',
+            url: 'https://sepolia-rollup.arbitrum.io/rpc',
             accounts: {
                 mnemonic: MNEMONIC,
                 initialIndex: 0,