+++ examples/oft/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @layerzerolabs/oft-example
 
+## 0.5.3
+
+### Patch Changes
+
+- 1d2abff: new SDK methods, tests in devtools-ton, upgraded lz-definitions
+
 ## 0.5.2
 
 ### Patch Changes