+++ packages/devtools-evm/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @layerzerolabs/devtools-evm
 
+## 1.0.2
+
+### Patch Changes
+
+- 1d2abff: new SDK methods, tests in devtools-ton, upgraded lz-definitions
+- Updated dependencies [1d2abff]
+  - @layerzerolabs/devtools@0.4.3
+
 ## 1.0.1
 
 ### Patch Changes