+++ packages/ua-devtools-evm/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @layerzerolabs/ua-devtools-evm
 
+## 5.0.3
+
+### Patch Changes
+
+- 1d2abff: new SDK methods, tests in devtools-ton, upgraded lz-definitions
+- Updated dependencies [1d2abff]
+  - @layerzerolabs/protocol-devtools-evm@3.0.3
+  - @layerzerolabs/protocol-devtools@1.1.2
+  - @layerzerolabs/devtools-evm@1.0.2
+  - @layerzerolabs/ua-devtools@3.0.2
+  - @layerzerolabs/devtools@0.4.3
+
 ## 5.0.2
 
 ### Patch Changes