+++ tests/devtools-evm-hardhat-test/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @layerzerolabs/devtools-evm-hardhat-test
 
+## 0.3.4
+
+### Patch Changes
+
+- 1d2abff: new SDK methods, tests in devtools-ton, upgraded lz-definitions
+
 ## 0.3.3
 
 ### Patch Changes